Version 1.2.1 Created a new activity called activity_selected_category. This activity will display that specific category that that user clicked on to see all the pictures of that specific clothing. As well as their is a camera for them to take more pictures.
Need to fix the recycler view so that when a user adds a category, they will be able to view their categories that they add.

Finally fixed the recycler view. User can add a category, and it will save it to the database as well as display it as a recycler view once they have added their category.