Version 1.2.2
Created an activity called activity_add_clothes_to_category_main. This activity will allow the user to take pictures of the type of clothing, i need to add a dropdown of the categories that they created as the drop down option. In the activity, they will be able to add the name, description and take a pciture. This will then save under the correct category of clothing type. If they take a pci of pants, then they must pick the category Pants so that it would be saved under that category.

Trying to create the back button for the activity_add_clothes_to_category_main. Still not working.