Version 1.2.3
I got images to be uploaded to the firebase storage. It displays the name of the image.
The problem i need to find a solution for is to upload image into a directory called images which will save the images there.

Later on i need to create a directory with images, then the user UID so that if a specific takes a picture, it would be stored under their ID.