Version 1.2.4
I left the images.

I got the app to only display the categories that they added. It won't display other users categories. It finally works. I added email to the addCategoryToFirebase method in the ClothesCategory class as well as to the ModelCategory and finally the AdapterCategory class.

Need to now do the same for the set goals.
*It displays in the database, but i cannot seem to display it in the recycler view. Still struggling.