Version 1.2.4
I added a new activity called activity_view_user_clothes. In this new activity, the user would be able to pick the category and whn they click search it should bring back the image, name and description of all the picks they took for that specific category.

Need to do the images.

Fix phone back button (Still need to do). The part where user logs out it takes them to the FrontCoverActivity, but when they click back...it takes them to the ClothesCategory page with the nav bar.