Version 1.2.4
User can add clothes (AddClothesToCategory class). It will save the users uid, id, name, description, category, url of the image and timestamp. It will show that in the realtime database and store the image under the storage section in firebase.

Still need to:
display the clothes depending on the user and the category.