Version 1.2.4
The user can add clothes by giving the name, description, choosing the category and taking a photo of the clothing. Once they click the upload button it will be uploaded. Then when they want to view their clothes, they must go to the View Clothes button in the nav bar and it will retrieve the images as well as the name and description of their clothes that they uploaded.

Need to create a search method. This method will allow user to pick a category from the dropdown menu and when they click search, it will bring back only the clothes that were placed under that specific category.